Implementing Connect Wallet Redesign (Part 1)
