+++ packages/app/components/constants.ts
@@ -14,6 +14,7 @@ import {
   ROPSTEN_TEST_TOKEN_ADDRESS,
 } from "@cabindao/nft-passport-contracts/artifacts/addresses";
 
+const MAINNET_NETWORK_ID = 0x1;
 const KOVAN_NETWORK_ID = 0x2a;
 const ROPSTEN_NETWORK_ID = 0x3;
 const RINKEBY_NETWORK_ID = 0x4;
@@ -26,6 +27,9 @@ export const networkNameById: {
   [LOCALHOST_NETWORK_ID]: "localhost",
   [KOVAN_NETWORK_ID]: "kovan",
   [ROPSTEN_NETWORK_ID]: "ropsten",
+  [MAINNET_NETWORK_ID]: "ethereum",
+  [RINKEBY_NETWORK_ID]: "rinkeby",
+  [GOERLI_NETWORK_ID]: "goerly"
 };
 
 export const networkIdByName = Object.fromEntries(
